pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "seraphs"
version = "2.0.0"
description = "Distributed Multi-Agent Compliance Intelligence System"
authors = [{name = "Seraphs Team"}]
requires-python = ">=3.11"
dependencies = [
    # Core Framework
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    
    # LangGraph & LangChain
    "langgraph>=0.0.40",
    "langchain>=0.1.0",
    "langchain-anthropic>=0.1.0",
    
    # Event Bus
    "redis>=5.0.0",
    
    # Storage
    "ipfshttpclient>=0.8.0a2",
    "psycopg2-binary>=2.9.9",
    
    # Web Fetching
    "requests>=2.31.0",
    "httpx>=0.25.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    
    # PDF Processing
    "PyPDF2>=3.0.0",
    "pdfplumber>=0.10.0",
    "pytesseract>=0.3.10",
    
    # RSS Parsing
    "feedparser>=6.0.10",
    
    # Scheduling
    "APScheduler>=3.10.0",
    
    # Security
    "cryptography>=41.0.0",
    
    # API Framework
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    
    # Logging
    "structlog>=23.2.0",

    # Production Enhancements
    "websockets>=12.0", # NEW: WebSocket support
    "sentence-transformers>=2.2.0", # NEW: Embeddings
    "torch>=2.1.0", # NEW: For transformers
    "numpy>=1.24.0", # NEW: Array operations
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
]

blockchain = [
    "pycardano>=0.9.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
